Finalização do Projeto :checkered_flag:
Término do Cálculo de Dano e modificações pontuais no Sistema de Batalha.

+++ N1/src/main/java/uam/n1/Elemonstro.java
@@ -3,13 +3,13 @@ package uam.n1;
 public class Elemonstro extends Golpe{
     private String nomeElemonstro;
     private String elementoElemonstro;
-    private int vida;
+    private double vida;
     private int forca;
     private int defesa;
     private int velocidade;
 
 
-    public Elemonstro(String nomeGolpe1, String nomeGolpe2, int poder1, int poder2, String elementoGolpe, String nomeElemonstro, String elementoElemonstro, int vida, int forca, int defesa, int velocidade) {
+    public Elemonstro(String nomeGolpe1, String nomeGolpe2, int poder1, int poder2, String elementoGolpe, String nomeElemonstro, String elementoElemonstro, double vida, int forca, int defesa, int velocidade) {
     super(nomeGolpe1,nomeGolpe2,poder1,poder2,elementoGolpe);
         this.nomeElemonstro = nomeElemonstro;
         this.elementoElemonstro = elementoElemonstro;
@@ -27,11 +27,11 @@ public class Elemonstro extends Golpe{
         return elementoElemonstro;
     }
 
-    public int getVida() {
+    public double getVida() {
         return vida;
     }
 
-    public void setVida(int vida) {
+    public void setVida(double vida) {
         this.vida = vida;
     }
